Refactor newBudgetForm into modular components for improved structure
- Deleted the original newBudgetForm.vue and created newBudget.vue and newBudgetForm.vue to enhance modularity and maintainability.
- Updated myBudgets.vue to utilize the new newBudget component, ensuring better organization and separation of concerns.
- Implemented responsive design with a Drawer for mobile views and a Dialog for desktop views, improving user experience across devices.
